+++ src/resources/views/widgets/avatar.blade.php
@@ -1,1 +1,2 @@
-<img src="{{ avatar_image_url($data, 100) }}" class="img-avatar img-avatar-50">
+<img src="{{ avatar_image_url($data, $size*2) }}" class="img-avatar img-avatar-{{ $size }}" style="width: {{ $size }}px;" @if($tooltip)title="{{ $tooltip }}"@endif>
+