+++ Changelog.md
@@ -6,11 +6,13 @@
 - Added actionbar and footer slots to the group widget
 - Added possibility to omit the title from group widget
 - Added "empty" option to table widget to render a text instead of table on empty dataset
+- Added "header:false" option to hide table headers
 - Added "value" option for getting a contextual color directly from the model
 - Added Footer support to the table widget
 - Added Raw HTML widget
 - Added support for conditional widget rendering
 - Added support for extra elements in multi text widget's primary row
+- Added size and tooltip rendering to Avatar widget
 - Changed customer and user show/edit/create pages to use group widget instead of hard coded Bootstrap
 - Changed widget text modifiers so that they can modify non-string field values as well
 - Converted address list partial to use the group widget instead of hardcoded bootstrap HTML